Renamed cytoscape package to editor
